Multiple changes to get PI Controller working on lift.

+++ src/Support/PIController.h
@@ -25,12 +25,12 @@ public:
 	double timeFilter;
 	double maxOutput;
 	double maxRate;
-	float target;
+	double target;
 	double lastTime;
 	double curTime;
 	double controlOutput;
 	bool rateController;
-	void SetTarget(float targetIn);
+	void SetTarget(double targetIn);
     void SetPosition(double output);
     void SetRate(double rate);
     PIController(double pGain, double iGain, double timeFilter, double maxOutputIn, double maxRateIn, double initPosition);
